+++ TSClientGen.Contract/Properties/AssemblyInfo.cs
@@ -1,8 +1,8 @@
 ﻿using System.Reflection;
 using System.Runtime.CompilerServices;
 
-[assembly: AssemblyVersion("5.2.0")]
-[assembly: AssemblyInformationalVersion("5.2.0")]
+[assembly: AssemblyVersion("5.3.0")]
+[assembly: AssemblyInformationalVersion("5.3.0")]
 
 [assembly: AssemblyTitle("TSClientGen.Contract")]
 [assembly: AssemblyDescription("Contracts assembly with attributes for decorating api for TSClientGen tool")]