+++ TSClientGen.NetFrameworkTool/AspNetWebApi/TypeConverter.cs
@@ -3,7 +3,7 @@ using System.Net.Http;
 using Newtonsoft.Json.Linq;
 using TSClientGen.Extensibility;
 
-namespace TSClientGen.AspNetWebApi
+namespace TSClientGen.NetFrameworkTool
 {
 	public class TypeConverter : ITypeConverter
 	{