+++ TSClientGen.NetFrameworkTool/AspNetWebApi/TypeDescriptorProvider.cs
@@ -5,7 +5,7 @@ using Newtonsoft.Json;
 using TSClientGen.Extensibility;
 using TSClientGen.Extensibility.ApiDescriptors;
 
-namespace TSClientGen.AspNetWebApi
+namespace TSClientGen.NetFrameworkTool
 {
 	public class TypeDescriptorProvider : ITypeDescriptorProvider
 	{