+++ TSClientGen.NetFrameworkTool/Properties/AssemblyInfo.cs
@@ -3,5 +3,5 @@
 [assembly: AssemblyVersion("5.3.0")]
 [assembly: AssemblyInformationalVersion("5.3.0")]
 
-[assembly: AssemblyTitle("TSClientGen.Tests")]
-[assembly: AssemblyDescription("Unit tests for TSClientGen tool")]
+[assembly: AssemblyTitle("TSClientGen.NetFrameworkTool")]
+[assembly: AssemblyDescription("")]